backend/requirements.txt
fastapi
uvicorn
python-dotenv
PyJWT
mysql-connector-python

